move TableScan-related stuff to new folder "table_scan", restructure source files
