+++ src/lib/CMakeLists.txt
@@ -9,8 +9,10 @@ set(
     operators/get_table.hpp
     operators/print.cpp
     operators/print.hpp
-    operators/table_scan.cpp
-    operators/table_scan.hpp
+    operators/table_scan/table_scan.cpp
+    operators/table_scan/table_scan.hpp
+    operators/table_scan/table_scan_impl.hpp
+    operators/table_scan/vector_scan.hpp
     operators/table_wrapper.cpp
     operators/table_wrapper.hpp
     storage/base_attribute_vector.hpp
