+++ src/lib/operators/table_scan/table_scan_impl.hpp
@@ -1,18 +1,11 @@
-#include "table_scan.hpp"
+#pragma once
 
-#include <functional>
-#include <limits>
 #include <memory>
 #include <vector>
 
-#include "resolve_type.hpp"
-#include "storage/table.hpp"
-
-#include "storage/base_column.hpp"
-#include "storage/dictionary_column.hpp"
-#include "storage/fitted_attribute_vector.hpp"
-#include "storage/reference_column.hpp"
-#include "storage/value_column.hpp"
+#include "table_scan.hpp"
+#include "types.hpp"
+#include "vector_scan.hpp"
 
 namespace opossum {
 
@@ -27,11 +20,11 @@ struct ReferenceGetter {
       : _table{table},
         _column_id{column_id},
         _is_last_chunk_valid{false},
-        _last_chunk_id{std::numeric_limits<ChunkID::base_type>::max()},
+        _last_chunk_id{0},
         _last_value_ptr{nullptr},
         _last_dictionary_ptr{nullptr} {}
   const T& operator()(const RowID& row_id) const {
-    if (! _is_last_chunk_valid || _last_chunk_id != row_id.chunk_id) {
+    if (!_is_last_chunk_valid || _last_chunk_id != row_id.chunk_id) {
       const Chunk& chunk = _table.get_chunk(row_id.chunk_id);
       const auto& column_ptr = chunk.get_column(_column_id);
 
@@ -39,7 +32,7 @@ struct ReferenceGetter {
       _last_dictionary_ptr = std::dynamic_pointer_cast<DictionaryColumn<T>>(column_ptr);
       _last_chunk_id = row_id.chunk_id;
 
-      _is_last_chunk_valid = true; // TODO is there an INVALID_CHUNK_ID?!
+      _is_last_chunk_valid = true;  // TODO is there an INVALID_CHUNK_ID?!
     }
 
     if (_last_value_ptr) {
@@ -72,59 +65,6 @@ struct ReferenceGetter {
   mutable std::shared_ptr<ValueColumn<T>> _last_value_ptr;
   mutable std::shared_ptr<DictionaryColumn<T>> _last_dictionary_ptr;
 };
-
-template <typename T, typename Getter, typename CompType>
-void vector_scan_impl(const std::vector<T>& values, const Getter& getter, const T& compare_value, PosList& pos_list,
-                      ChunkID chunk_id) {
-  CompType comparator;
-  for (ChunkOffset chunk_offset{0}; chunk_offset < values.size(); ++chunk_offset) {
-    const auto& value = getter(values[chunk_offset]);
-    if (comparator(value, compare_value)) {
-      pos_list.emplace_back(RowID{chunk_id, chunk_offset});
-    }
-  }
-}
-
-template <typename T, typename Getter, typename CompType>
-void vector_scan_impl(const std::vector<RowID>& values, const Getter& getter, const T& compare_value, PosList& pos_list,
-                      ChunkID chunk_id) {
-  CompType comparator;
-  for (const RowID& row_id : values) {
-    const auto& value = getter(row_id);
-    if (comparator(value, compare_value)) {
-      pos_list.emplace_back(row_id);
-    }
-  }
-}
-
-template <typename VectorT, typename CompareT, typename Getter>
-void vector_scan(const std::vector<VectorT>& values, const Getter& getter, const CompareT& compare_value,
-                 PosList& pos_list, ChunkID chunk_id, ScanType scan_type) {
-  switch (scan_type) {
-    case ScanType::OpEquals:
-      vector_scan_impl<CompareT, Getter, std::equal_to<CompareT>>(values, getter, compare_value, pos_list, chunk_id);
-      break;
-    case ScanType::OpNotEquals:
-      vector_scan_impl<CompareT, Getter, std::not_equal_to<CompareT>>(values, getter, compare_value, pos_list,
-                                                                      chunk_id);
-      break;
-    case ScanType::OpLessThan:
-      vector_scan_impl<CompareT, Getter, std::less<CompareT>>(values, getter, compare_value, pos_list, chunk_id);
-      break;
-    case ScanType::OpLessThanEquals:
-      vector_scan_impl<CompareT, Getter, std::less_equal<CompareT>>(values, getter, compare_value, pos_list, chunk_id);
-      break;
-    case ScanType::OpGreaterThan:
-      vector_scan_impl<CompareT, Getter, std::greater<CompareT>>(values, getter, compare_value, pos_list, chunk_id);
-      break;
-    case ScanType::OpGreaterThanEquals:
-      vector_scan_impl<CompareT, Getter, std::greater_equal<CompareT>>(values, getter, compare_value, pos_list,
-                                                                       chunk_id);
-      break;
-    default:
-      throw std::runtime_error("Operator not known");
-  }
-}
 
 class BaseTableScanImpl {
  public:
@@ -295,54 +235,5 @@ class TypedTableScanImpl : public BaseTableScanImpl {
   const T _search_value;
   const ScanType _scan_type;
 };
-
-TableScan::TableScan(const std::shared_ptr<const AbstractOperator> in, ColumnID column_id, const ScanType scan_type,
-                     const AllTypeVariant search_value)
-    : AbstractOperator(in), _column_id{column_id}, _scan_type{scan_type}, _search_value{search_value} {}
-
-ColumnID TableScan::column_id() const { return _column_id; }
-
-ScanType TableScan::scan_type() const { return _scan_type; }
-
-const AllTypeVariant& TableScan::search_value() const { return _search_value; }
-
-std::shared_ptr<const Table> TableScan::_on_execute() {
-  auto input_table = _input_table_left();
-  Assert(input_table, "No output");
-  const auto& column_type = input_table->column_type(_column_id);
-  auto table_scan_impl = make_shared_by_column_type<BaseTableScanImpl, TypedTableScanImpl>(column_type, *this);
-
-  auto pos_list_ptr = table_scan_impl->on_execute();
-
-  // We assume:
-  // 1. A table either consists of ReferenceColumns (X)OR Value- or DictionaryColumns.
-  // 2. All ReferenceColumns in a table reference the same "original" table.
-  // Therefore, we only need to handle one single PosList to create our new Table / ReferenceColumn.
-
-  // If the input table consists of ReferenceColumns, take the original reference to avoid multiple indirections
-  auto referenced_table = input_table;
-  auto input_reference_column_ptr =
-      std::dynamic_pointer_cast<ReferenceColumn>(input_table->get_chunk(ChunkID{0}).get_column(_column_id));
-  if (input_reference_column_ptr) {
-    referenced_table = input_reference_column_ptr->referenced_table();
-  }
-
-  auto output_table = std::make_shared<Table>();
-
-  // Copy column definitions from referenced table
-  for (ColumnID column_id{0}; column_id < referenced_table->col_count(); ++column_id) {
-    output_table->add_column_definition(referenced_table->column_name(column_id),
-                                        referenced_table->column_type(column_id));
-  }
-
-  // Add ReferenceColumns to first chunk (all using the same PosList)
-  auto& chunk = output_table->get_chunk(ChunkID{0});
-  for (ColumnID column_id{0}; column_id < referenced_table->col_count(); ++column_id) {
-    auto reference_column = std::make_shared<ReferenceColumn>(referenced_table, column_id, pos_list_ptr);
-    chunk.add_column(reference_column);
-  }
-
-  return output_table;
-}
 
 }  // namespace opossum
